SDL: rework window resizing
Only change the font when the window is resized, this:
   1/ Maintains aspect ratio (as before) but instead of resizing the
window to fit the new cell size addis offsets to centre the drawn text
within the user defined window size.
   2/ Sometimes the calculated font size is too large for the user
defined window size resulting in clipped edges of the text. This is
because scaling the pointsize is not truly linear, at least with the
default font. Possibly a better calcuation could be obtained by
iterating through a range of pointsizes until suitable match is found.
That seems to be an expensize operation on SDL_TTF as the font has to be
opened at each pointsize to glean the respective height/width.